feat(language): VSCODE-24: Language server setup (#47)
This PR provides the initial setup for MongoDB Language Server; primarily, it is documentation, dependencies and test fixture setup we'll add user-facing features on top of in the future. I also added a static mongodb language configuration in ./language for the editor.

A test fixture workspace has been added under `src/test/fixture`.