chore: infer MongoClientOptions type instead of importing it
Rather than importing it, define it in a slightly more reliable way
by referring to the second argument of `CliServiceProvider.connect()`.

For example, this also enables usage of `tlsUseSystemCA`, which is
included in `DevtoolsConnectOptions` (which is now the actual type
accepted by `CliServiceProvider.connect()`) but not in the driver's
`MongoClientOptions`.