fix: resolve SRV hostname before passing it to mongodb-cloud-info VSCODE-442
Right now, we are passing the first hostname from the connection string
to mongodb-cloud-info, regardless of whether the connection string is a
`mongodb://` connection string or a `mongodb+srv://` one.

In the latter case, telemetry gathering fails, because
mongodb-cloud-info expects the name of an actual host with associated
A or AAAA records (so that it can look up the host’s IP address),
not a SRV one.

Using resolve-mongodb-srv first matches what Compass does and
solves this issue.