Skunktendo 2024: automatically detect Local Atlas
With this PR, the VS Code extension detects local atlas environments
and shows them in the list of connections so that the user does not
need to pass connection strings around.

It does so by looking at the running docker containers, as that is
how local atlas environments are provided.

As added bonus, connections can now be open in Compass directy from VSCode
if the protocol handler is registered.