chore: modifies server activation flow
Earlier we were starting the server only after a connection was
connected for the first time and then additionally showing notification
to disable the auto connect and auto start.

With this change, we will start the server as soon as the extension
activates and will show a notification to the user that auto-start of
MongoDB MCP server has been enabled. If they wish to disable this
behavior, the popup has buttons to do that. If they choose to disable
then we will stop the server and set the auto-start to disable.