chore: disable MCP auto start and instead request
This commit flips the earlier behaviour of MongoDB MCP server
auto-starting on VSCode activate. We will now popup a notification
asking user to enable auto-start or disable it. If the user does not
select any option or the notification auto-dimisses then we will ask for
the same on every start of VSCode and every new connection attempt.