chore: reorganise GA to check and build once
Earlier we were running check, build and tests on each OS in our matrix
which did not allow us to test whether a build on ubuntu would work fine
on Windows / MacOS or not.

With this commit we are re-organising to check and build once on ubuntu
and use the final artifact later in the test job to run tests against.

Accordingly the test-and-build-from-fork and draft-release workflows
have been modified to adapt the same change.